Broke the different busses into different modules. Added more examples. Cleaned up code. Working on adding documentation.
