+++ Cargo.toml
@@ -4,26 +4,21 @@ name = "icm20948_driver"
 edition = "2018"
 version = "0.1.0"
 
-[workspace]
-members = ["testsuite"]
-
-[features]
-default = ["i2c"]
-i2c = []
-
 [dependencies]
+embedded-hal = "0.2.7"
 cortex-m = "0.7"
 defmt = "0.3.0"
 defmt-rtt = "0.3.0"
 panic-probe = { version = "0.3.0", features = ["print-defmt"] }
 
-[dependencies.stm32h7xx-hal]
-version = "0.12.2"
-features = ["stm32h743v"]
-
 [dev-dependencies]
 cortex-m-rtic = "1.1"
 systick-monotonic = "1.0.0"
+cortex-m-rt = "0.7.0"
+
+[dev-dependencies.stm32h7xx-hal]
+version = "0.12.2"
+features = ["stm32h743v"]
 
 # cargo build/run
 [profile.dev]
