--- testsuite/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-authors = ["zolkin"]
-name = "testsuite"
-publish = false
-edition = "2021"
-version = "0.1.0"
-
-[features]
-
-[lib]
-harness = false
-
-[[test]]
-name = "test"
-harness = false
-
-[dependencies]
-icm20948_driver = { path = ".." }
-cortex-m = "0.7"
-#cortex-m-rt = "0.7"
-defmt = "0.3.0"
-defmt-rtt = "0.3.0"
-defmt-test = "0.3.0"
-panic-probe = { version = "0.3.0", features = ["print-defmt"] }
-
-[dependencies.stm32h7xx-hal]
-version = "0.12.2"
-features = ["stm32h743v"]
